Avoid module inception in layout.rs
