+++ src/layout/spring_layout.rs
@@ -10,18 +10,24 @@
 // License for the specific language governing permissions and limitations
 // under the License.
 
+use crate::iterators::Pos2DMapping;
+use crate::weight_callable;
+
 use std::iter::Iterator;
 
 use hashbrown::{HashMap, HashSet};
 
+use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 
 use petgraph::graph::NodeIndex;
 use petgraph::prelude::*;
-use petgraph::visit::NodeIndexable;
+use petgraph::visit::{IntoEdgeReferences, NodeIndexable};
 use petgraph::EdgeType;
 
-use crate::iterators::Pos2DMapping;
+use rand::distributions::{Distribution, Uniform};
+use rand::prelude::*;
+use rand_pcg::Pcg64;
 
 type Nt = f64;
 pub type Point = [Nt; 2];
@@ -162,7 +168,7 @@ impl CoolingScheme for LinearCoolingScheme {
 }
 
 // Rescale so that pos in [-scale, scale].
-fn rescale(pos: &mut Vec<Point>, scale: Nt, indices: Vec<usize>) {
+pub fn rescale(pos: &mut Vec<Point>, scale: Nt, indices: Vec<usize>) {
     let n = indices.len();
     if n == 0 {
         return;
@@ -198,7 +204,7 @@ fn rescale(pos: &mut Vec<Point>, scale: Nt, indices: Vec<usize>) {
     }
 }
 
-fn recenter(pos: &mut Vec<Point>, center: Point) {
+pub fn recenter(pos: &mut Vec<Point>, center: Point) {
     for [px, py] in pos.iter_mut() {
         *px += center[0];
         *py += center[1];
@@ -290,167 +296,83 @@ where
     pos
 }
 
-pub fn bipartite_layout<Ty: EdgeType>(
+#[allow(clippy::too_many_arguments)]
+pub fn spring_layout<Ty>(
+    py: Python,
     graph: &StableGraph<PyObject, PyObject, Ty>,
-    first_nodes: HashSet<usize>,
-    horizontal: Option<bool>,
+    pos: Option<HashMap<usize, Point>>,
+    fixed: Option<HashSet<usize>>,
+    k: Option<f64>,
+    repulsive_exponent: Option<i32>,
+    adaptive_cooling: Option<bool>,
+    num_iter: Option<usize>,
+    tol: Option<f64>,
+    weight_fn: Option<PyObject>,
+    default_weight: f64,
     scale: Option<f64>,
     center: Option<Point>,
-    aspect_ratio: Option<f64>,
-) -> Pos2DMapping {
-    let node_num = graph.node_count();
-    if node_num == 0 {
-        return Pos2DMapping {
-            pos_map: HashMap::new(),
-        };
-    }
-    let left_num = first_nodes.len();
-    let right_num = node_num - left_num;
-    let mut pos: Vec<Point> = Vec::with_capacity(node_num);
-
-    let (width, height);
-    if horizontal == Some(true) {
-        // width and height viewed from 90 degrees clockwise rotation
-        width = 1.0;
-        height = match aspect_ratio {
-            Some(aspect_ratio) => aspect_ratio * width,
-            None => 4.0 * width / 3.0,
-        };
-    } else {
-        height = 1.0;
-        width = match aspect_ratio {
-            Some(aspect_ratio) => aspect_ratio * height,
-            None => 4.0 * height / 3.0,
-        };
+    seed: Option<u64>,
+) -> PyResult<Pos2DMapping>
+where
+    Ty: EdgeType,
+{
+    if fixed.is_some() && pos.is_none() {
+        return Err(PyValueError::new_err("`fixed` specified but `pos` not."));
     }
 
-    let x_offset: f64 = width / 2.0;
-    let y_offset: f64 = height / 2.0;
-    let left_dy: f64 = match left_num {
-        0 | 1 => 0.0,
-        _ => height / (left_num - 1) as f64,
-    };
-    let right_dy: f64 = match right_num {
-        0 | 1 => 0.0,
-        _ => height / (right_num - 1) as f64,
+    let mut rng: Pcg64 = match seed {
+        Some(seed) => Pcg64::seed_from_u64(seed),
+        None => Pcg64::from_entropy(),
     };
 
-    let mut lc: f64 = 0.0;
-    let mut rc: f64 = 0.0;
+    let dist = Uniform::new(0.0, 1.0);
 
-    for node in graph.node_indices() {
-        let n = node.index();
+    let pos = pos.unwrap_or_default();
+    let mut vpos: Vec<Point> = (0..graph.node_bound())
+        .map(|_| [dist.sample(&mut rng), dist.sample(&mut rng)])
+        .collect();
+    for (n, p) in pos.into_iter() {
+        vpos[n] = p;
+    }
 
-        let (x, y): (f64, f64);
-        if first_nodes.contains(&n) {
-            x = -x_offset;
-            y = lc * left_dy - y_offset;
-            lc += 1.0;
-        } else {
-            x = width - x_offset;
-            y = rc * right_dy - y_offset;
-            rc += 1.0;
+    let fixed = fixed.unwrap_or_default();
+    let k = k.unwrap_or(1.0 / (graph.node_count() as f64).sqrt());
+    let f_a = AttractiveForce::new(k);
+    let f_r = RepulsiveForce::new(k, repulsive_exponent.unwrap_or(2));
+
+    let num_iter = num_iter.unwrap_or(50);
+    let tol = tol.unwrap_or(1e-6);
+    let step = 0.1;
+
+    let mut weights: HashMap<(usize, usize), f64> =
+        HashMap::with_capacity(2 * graph.edge_count());
+    for e in graph.edge_references() {
+        let w = weight_callable(py, &weight_fn, e.weight(), default_weight)?;
+        let source = e.source().index();
+        let target = e.target().index();
+
+        weights.insert((source, target), w);
+        weights.insert((target, source), w);
+    }
+
+    let pos = match adaptive_cooling {
+        Some(false) => {
+            let cs = LinearCoolingScheme::new(step, num_iter);
+            evolve(
+                graph, vpos, fixed, f_a, f_r, cs, num_iter, tol, weights,
+                scale, center,
+            )
         }
-
-        if horizontal == Some(true) {
-            pos.push([-y, x]);
-        } else {
-            pos.push([x, y]);
+        _ => {
+            let cs = AdaptiveCoolingScheme::new(step);
+            evolve(
+                graph, vpos, fixed, f_a, f_r, cs, num_iter, tol, weights,
+                scale, center,
+            )
         }
-    }
-
-    if let Some(scale) = scale {
-        rescale(&mut pos, scale, (0..node_num).collect());
-    }
-
-    if let Some(center) = center {
-        recenter(&mut pos, center);
-    }
-
-    Pos2DMapping {
-        pos_map: graph.node_indices().map(|n| n.index()).zip(pos).collect(),
-    }
-}
-
-pub fn circular_layout<Ty: EdgeType>(
-    graph: &StableGraph<PyObject, PyObject, Ty>,
-    scale: Option<f64>,
-    center: Option<Point>,
-) -> Pos2DMapping {
-    let node_num = graph.node_count();
-    let mut pos: Vec<Point> = Vec::with_capacity(node_num);
-    let pi = std::f64::consts::PI;
-
-    if node_num == 1 {
-        pos.push([0.0, 0.0])
-    } else {
-        for i in 0..node_num {
-            let angle = 2.0 * pi * i as f64 / node_num as f64;
-            pos.push([angle.cos(), angle.sin()]);
-        }
-    }
-
-    if let Some(scale) = scale {
-        rescale(&mut pos, scale, (0..node_num).collect());
-    }
-
-    if let Some(center) = center {
-        recenter(&mut pos, center);
-    }
-
-    Pos2DMapping {
-        pos_map: graph.node_indices().map(|n| n.index()).zip(pos).collect(),
-    }
-}
-
-pub fn shell_layout<Ty: EdgeType>(
-    graph: &StableGraph<PyObject, PyObject, Ty>,
-    nlist: Option<Vec<Vec<usize>>>,
-    rotate: Option<f64>,
-    scale: Option<f64>,
-    center: Option<Point>,
-) -> Pos2DMapping {
-    let node_num = graph.node_bound();
-    let mut pos: Vec<Point> = vec![[0.0, 0.0]; node_num];
-    let pi = std::f64::consts::PI;
-
-    let shell_list: Vec<Vec<usize>> = match nlist {
-        Some(nlist) => nlist,
-        None => vec![graph.node_indices().map(|n| n.index()).collect()],
-    };
-    let shell_num = shell_list.len();
-
-    let radius_bump = match scale {
-        Some(scale) => scale / shell_num as f64,
-        None => 1.0 / shell_num as f64,
     };
 
-    let mut radius = match node_num {
-        1 => 0.0,
-        _ => radius_bump,
-    };
-
-    let rot_angle = match rotate {
-        Some(rotate) => rotate,
-        None => pi / shell_num as f64,
-    };
-
-    let mut first_theta = rot_angle;
-    for shell in shell_list {
-        let shell_len = shell.len();
-        for i in 0..shell_len {
-            let angle = 2.0 * pi * i as f64 / shell_len as f64 + first_theta;
-            pos[shell[i]] = [radius * angle.cos(), radius * angle.sin()];
-        }
-        radius += radius_bump;
-        first_theta += rot_angle;
-    }
-
-    if let Some(center) = center {
-        recenter(&mut pos, center);
-    }
-
-    Pos2DMapping {
+    Ok(Pos2DMapping {
         pos_map: graph
             .node_indices()
             .map(|n| {
@@ -458,52 +380,5 @@ pub fn shell_layout<Ty: EdgeType>(
                 (n, pos[n])
             })
             .collect(),
-    }
-}
-
-pub fn spiral_layout<Ty: EdgeType>(
-    graph: &StableGraph<PyObject, PyObject, Ty>,
-    scale: Option<f64>,
-    center: Option<Point>,
-    resolution: Option<f64>,
-    equidistant: Option<bool>,
-) -> Pos2DMapping {
-    let node_num = graph.node_count();
-    let mut pos: Vec<Point> = Vec::with_capacity(node_num);
-
-    let ros = resolution.unwrap_or(0.35);
-
-    if node_num == 1 {
-        pos.push([0.0, 0.0]);
-    } else if equidistant == Some(true) {
-        let mut theta: f64 = ros;
-        let chord = 1.0;
-        let step = 0.5;
-        for _ in 0..node_num {
-            let r = step * theta;
-            theta += chord / r;
-            pos.push([theta.cos() * r, theta.sin() * r]);
-        }
-    } else {
-        let mut angle: f64 = 0.0;
-        let mut dist = 0.0;
-        let step = 1.0;
-        for _ in 0..node_num {
-            pos.push([dist * angle.cos(), dist * angle.sin()]);
-            dist += step;
-            angle += ros;
-        }
-    }
-
-    if let Some(scale) = scale {
-        rescale(&mut pos, scale, (0..node_num).collect());
-    }
-
-    if let Some(center) = center {
-        recenter(&mut pos, center);
-    }
-
-    Pos2DMapping {
-        pos_map: graph.node_indices().map(|n| n.index()).zip(pos).collect(),
-    }
+    })
 }
